backend/requirements.txt
asgiref==3.7.2
black==23.12.1
certifi==2023.11.17
cfgv==3.4.0
charset-normalizer==3.3.2
click==8.1.7
distlib==0.3.8
Django==5.0.1
django-cors-headers==4.3.1
djangorestframework==3.14.0
filelock==3.13.1
identify==2.5.33
idna==3.6
mypy-extensions==1.0.0
nodeenv==1.8.0
packaging==23.2
pathspec==0.12.1
platformdirs==4.1.0
pre-commit==3.6.0
psycopg2==2.9.9
psycopg2-binary==2.9.9
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
sqlparse==0.4.4
urllib3==2.1.0
virtualenv==20.25.0
